Move non-middle-wheel commands to own package
