+++ src/main/java/frc/team4373/robot/commands/auton/drive/fallback/NonMiddleWheelAdjustmentGroup.java
@@ -1,7 +1,9 @@
-package frc.team4373.robot.commands.auton.drive;
+package frc.team4373.robot.commands.auton.drive.fallback;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import frc.team4373.robot.RobotMap;
+import frc.team4373.robot.commands.auton.drive.DriveDistanceAuton;
+import frc.team4373.robot.commands.auton.drive.TurnToAngleAuton;
 
 public class NonMiddleWheelAdjustmentGroup extends CommandGroup {
     /**
