+++ src/main/java/frc/team4373/robot/commands/auton/drive/fallback/NonMiddleWheelLateralAdjusterAuton.java
@@ -1,7 +1,8 @@
-package frc.team4373.robot.commands.auton.drive;
+package frc.team4373.robot.commands.auton.drive.fallback;
 
 import edu.wpi.first.wpilibj.command.Scheduler;
 import frc.team4373.robot.RobotMap;
+import frc.team4373.robot.commands.auton.drive.VisionCommand;
 
 public class NonMiddleWheelLateralAdjusterAuton extends VisionCommand {
