+++ src/main/java/frc/team4373/robot/commands/auton/drive/fallback/VisionRotatorAuton.java
@@ -1,7 +1,9 @@
-package frc.team4373.robot.commands.auton.drive;
+package frc.team4373.robot.commands.auton.drive.fallback;
 
 import edu.wpi.first.wpilibj.command.Scheduler;
 import frc.team4373.robot.RobotMap;
+import frc.team4373.robot.commands.auton.drive.TurnToAngleAuton;
+import frc.team4373.robot.commands.auton.drive.VisionCommand;
 import frc.team4373.robot.subsystems.Drivetrain;
 
 public class VisionRotatorAuton extends VisionCommand {
